Performance optimization with Context API
Utilize Context API instead of props drilling so the entire VDOM tree doesn't need to re-render when the states in the App component changed
