+++ src/components/Scoreboard.jsx
@@ -1,6 +1,9 @@
-import PropTypes from 'prop-types'
+import { useContext } from 'react'
+import { Context } from '../context/ContextProvider'
 
-function Scoreboard({ score, bestScore }) {
+function Scoreboard() {
+  const { score, bestScore } = useContext(Context)
+
   return (
     <div className='rounded-lg bg-black bg-opacity-50 p-4 text-3xl font-bold'>
       <p>Current Score: {score}</p>
@@ -15,11 +18,6 @@ export function Instruction() {
       <p>Don&apos;t click the same Pokémon twice!</p>
     </div>
   )
-}
-
-Scoreboard.propTypes = {
-  score: PropTypes.number.isRequired,
-  bestScore: PropTypes.number.isRequired
 }
 
 export default Scoreboard
